Added metrics capture for app level metrics
